addressing comments PR#1996
