+++ openpype/lib/__init__.py
@@ -144,7 +144,8 @@ from .path_tools import (
     version_up,
     get_version_from_path,
     get_last_version_from_path,
-    create_project_folders
+    create_project_folders,
+    get_project_basic_paths
 )
 
 from .editorial import (
@@ -278,5 +279,6 @@ __all__ = [
     "frames_to_secons",
     "frames_to_timecode",
     "make_sequence_collection",
-    "create_project_folders"
+    "create_project_folders",
+    "get_project_basic_paths"
 ]
