+++ openpype/modules/default_modules/ftrack/event_handlers_user/action_create_project_structure.py
@@ -2,7 +2,7 @@ import os
 import re
 import json
 
-from openpype.modules.ftrack.lib import BaseAction, statics_icon
+from openpype_modules.ftrack.lib import BaseAction, statics_icon
 from openpype.api import get_project_basic_paths, create_project_folders
